moved docs generate and parse methods into seperate files
